Moved CachePeers class into a separate header
Also adjusted method/function names, iterator constness
and documentation.
